Update README with examples for v0.1.0 release
Include basic examples showing the expected syntax.

Because of the behavior noted in GH-12 (implemented in GH-3),
some of the examples showcase what may be unexpected results.
In short, a WARNING threshold may be triggered before a CRITICAL
state threshold. This will require further work to achieve a
balance of both goals, but won't make into the v0.1.0 release
scheduled for today.

refs GH-9
refs GH-12, GH-3