Add fail-fast flag, disable behavior by default
While the benefit of quickly failing a large collection
of content is that the service check is less likely to
timeout, the downside is that the state result from
using this behavior is becomes indeterminate.

When using fail-fast logic, a WARNING state result can
be returned for a collection of content where a CRITICAL
state is more appropriate. This could be potentially
confusing when troubleshooting a system flagged by a
failing check.

To help prevent confusion, the fail-fast logic has been
made optional, disabled by default. To use it, a flag
of the same name has been provided to re-enable the
behavior as desired.

NOTE: This commit does not include updated usage examples
to reflect the behavior change. This will follow shortly.

This commit also includes some unrelated changes mixed in
while chasing rabbits; I've reduced some duplication and
tweaked code formatting in places near relevant code changes.

rfs GH-12