Refactor checks, fix fail-fast logic
If specified, the fail-fast logic is applied to items as they are
encountered, otherwise a collection of all encountered items is
built and evaluated after.

As part of this work a number of refactoring steps were applied to
reduce the complexity of the main function and isolate specific
check code within applicable functions.

New helper/wrapper types and a getter method were added to make
it easier to pass related collections of values and error
messages extracted out to the paths package to enable code reuse
and potential comparison later.

New methods for sorting collections by size were also added.

refs GH-33