Update vuln analysis GHAW to use on.push hook
This hook is needed for proper operation of the
`Vulnerability / CodeQL` job so that it can compare
before/after changes against the base branch.

I'm opting to skip limiting either of the on.push or the
on.pull_request hook events to just the base branch, instead
adding a commented directive to imply that I explicitly
made that decision.

To keep *all* jobs from running again on push events in
addition to the pull request events we check the event
type in the job definition and skip running the job
unless it is not a push event.

refs atc0005/todo#56